+++ index.html
@@ -467,7 +467,6 @@
       </div>
     </template>
 
-    <script src="js/main.js"></script>
-    <script src="js/functions.js"></script>
+    <script src="js/main.js" type="module"></script>
   </body>
 </html>
